# src/main/java/org/sahil/embl/person/model/PersonDetails.java
package org.sahil.embl.person.model;


import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonProperty;

@JsonIgnoreProperties(ignoreUnknown = true)
public class PersonDetails extends Person {


    @JsonProperty("personId")
    private Long personId;

    /**
     * @return personId
     **/
    public Long getPersonId() {
        return personId;
    }

    public void setPersonId(Long personId) {
        this.personId = personId;
    }
}